+++ packages/ckeditor5-clipboard/src/dragdrop.ts
@@ -7,32 +7,46 @@
  * @module clipboard/dragdrop
  */
 
-import { Plugin, type Editor } from '@ckeditor/ckeditor5-core';
+import { Plugin } from '@ckeditor/ckeditor5-core';
 
 import {
 	LiveRange,
 	MouseObserver,
 	type DataTransfer,
 	type Element,
-	type Position,
+	type Model,
 	type Range,
+	type Position,
 	type ViewDocumentMouseDownEvent,
 	type ViewDocumentMouseUpEvent,
 	type ViewElement,
-	type ViewRange
+	type DomEventData
 } from '@ckeditor/ckeditor5-engine';
 
-import { Widget, isWidget, type WidgetToolbarRepository } from '@ckeditor/ckeditor5-widget';
+import {
+	Widget,
+	isWidget,
+	type WidgetToolbarRepository
+} from '@ckeditor/ckeditor5-widget';
 
 import {
 	env,
 	uid,
+	global,
+	createElement,
+	DomEmitterMixin,
 	delay,
+	Rect,
 	type DelayedFunc,
-	type ObservableChangeEvent
+	type ObservableChangeEvent,
+	type DomEmitter
 } from '@ckeditor/ckeditor5-utils';
 
-import ClipboardPipeline, { type ClipboardContentInsertionEvent, type ViewDocumentClipboardOutputEvent } from './clipboardpipeline';
+import ClipboardPipeline, {
+	type ClipboardContentInsertionEvent,
+	type ViewDocumentClipboardOutputEvent
+} from './clipboardpipeline';
+
 import ClipboardObserver, {
 	type ViewDocumentDragEndEvent,
 	type ViewDocumentDragEnterEvent,
@@ -42,7 +56,7 @@ import ClipboardObserver, {
 	type ViewDocumentClipboardInputEvent
 } from './clipboardobserver';
 
-import { throttle, type DebouncedFunc } from 'lodash-es';
+import DragDropTarget from './dragdroptarget';
 
 import '../theme/clipboard.css';
 
@@ -128,6 +142,8 @@ import '../theme/clipboard.css';
  * The drag and drop feature. It works on top of the {@link module:clipboard/clipboardpipeline~ClipboardPipeline}.
  *
  * Read more about the clipboard integration in the {@glink framework/deep-dive/clipboard clipboard deep-dive} guide.
+ *
+ * @internal
  */
 export default class DragDrop extends Plugin {
 	/**
@@ -148,19 +164,26 @@ export default class DragDrop extends Plugin {
 	private _draggableElement!: Element | null;
 
 	/**
-	 * A throttled callback updating the drop marker.
-	 */
-	private _updateDropMarkerThrottled!: DebouncedFunc<( targetRange: Range ) => void>;
-
-	/**
-	 * A delayed callback removing the drop marker.
-	 */
-	private _removeDropMarkerDelayed!: DelayedFunc<() => void>;
-
-	/**
 	 * A delayed callback removing draggable attributes.
 	 */
-	private _clearDraggableAttributesDelayed!: DelayedFunc<() => void>;
+	private _clearDraggableAttributesDelayed: DelayedFunc<() => void> = delay( () => this._clearDraggableAttributes(), 40 );
+
+	/**
+	 * Whether the dragged content can be dropped only in block context.
+	 */
+	// TODO handle drag from other editor instance
+	// TODO configure to use block, inline or both
+	private _blockMode: boolean = false;
+
+	/**
+	 * DOM Emitter.
+	 */
+	private _domEmitter: DomEmitter = new ( DomEmitterMixin() )();
+
+	/**
+	 * The DOM element used to generate dragged preview image.
+	 */
+	private _previewContainer?: HTMLElement;
 
 	/**
 	 * @inheritDoc
@@ -173,7 +196,7 @@ export default class DragDrop extends Plugin {
 	 * @inheritDoc
 	 */
 	public static get requires() {
-		return [ ClipboardPipeline, Widget ] as const;
+		return [ ClipboardPipeline, Widget, DragDropTarget ] as const;
 	}
 
 	/**
@@ -186,15 +209,6 @@ export default class DragDrop extends Plugin {
 		this._draggedRange = null;
 		this._draggingUid = '';
 		this._draggableElement = null;
-		this._updateDropMarkerThrottled = throttle( targetRange => this._updateDropMarker( targetRange ), 40 );
-		this._removeDropMarkerDelayed = delay( () => this._removeDropMarker(), 40 );
-		this._clearDraggableAttributesDelayed = delay( () => this._clearDraggableAttributes(), 40 );
-
-		if ( editor.plugins.has( 'DragDropExperimental' ) ) {
-			this.forceDisabled( 'DragDropExperimental' );
-
-			return;
-		}
 
 		view.addObserver( ClipboardObserver );
 		view.addObserver( MouseObserver );
@@ -202,7 +216,6 @@ export default class DragDrop extends Plugin {
 		this._setupDragging();
 		this._setupContentInsertionIntegration();
 		this._setupClipboardInputIntegration();
-		this._setupDropMarker();
 		this._setupDraggableAttributeHandling();
 
 		this.listenTo<ObservableChangeEvent<boolean>>( editor, 'change:isReadOnly', ( evt, name, isReadOnly ) => {
@@ -233,8 +246,11 @@ export default class DragDrop extends Plugin {
 			this._draggedRange = null;
 		}
 
-		this._updateDropMarkerThrottled.cancel();
-		this._removeDropMarkerDelayed.cancel();
+		if ( this._previewContainer ) {
+			this._previewContainer.remove();
+		}
+
+		this._domEmitter.stopListening();
 		this._clearDraggableAttributesDelayed.cancel();
 
 		return super.destroy();
@@ -246,14 +262,12 @@ export default class DragDrop extends Plugin {
 	private _setupDragging(): void {
 		const editor = this.editor;
 		const model = editor.model;
-		const modelDocument = model.document;
 		const view = editor.editing.view;
 		const viewDocument = view.document;
+		const dragDropTarget = editor.plugins.get( DragDropTarget );
 
 		// The handler for the drag start; it is responsible for setting data transfer object.
 		this.listenTo<ViewDocumentDragStartEvent>( viewDocument, 'dragstart', ( evt, data ) => {
-			const selection = modelDocument.selection;
-
 			// Don't drag the editable element itself.
 			if ( data.target && data.target.is( 'editableElement' ) ) {
 				data.preventDefault();
@@ -261,34 +275,7 @@ export default class DragDrop extends Plugin {
 				return;
 			}
 
-			// TODO we could clone this node somewhere and style it to match editing view but without handles,
-			//  selection outline, WTA buttons, etc.
-			// data.dataTransfer._native.setDragImage( data.domTarget, 0, 0 );
-
-			// Check if this is dragstart over the widget (but not a nested editable).
-			const draggableWidget = data.target ? findDraggableWidget( data.target ) : null;
-
-			if ( draggableWidget ) {
-				const modelElement = editor.editing.mapper.toModelElement( draggableWidget )!;
-
-				this._draggedRange = LiveRange.fromRange( model.createRangeOn( modelElement ) );
-
-				// Disable toolbars so they won't obscure the drop area.
-				if ( editor.plugins.has( 'WidgetToolbarRepository' ) ) {
-					const widgetToolbarRepository: WidgetToolbarRepository = editor.plugins.get( 'WidgetToolbarRepository' );
-
-					widgetToolbarRepository.forceDisabled( 'dragDrop' );
-				}
-			}
-
-			// If this was not a widget we should check if we need to drag some text content.
-			else if ( !viewDocument.selection.isCollapsed ) {
-				const selectedElement = viewDocument.selection.getSelectedElement();
-
-				if ( !selectedElement || !isWidget( selectedElement ) ) {
-					this._draggedRange = LiveRange.fromRange( selection.getFirstRange()! );
-				}
-			}
+			this._prepareDraggedRange( data.target );
 
 			if ( !this._draggedRange ) {
 				data.preventDefault();
@@ -298,9 +285,7 @@ export default class DragDrop extends Plugin {
 
 			this._draggingUid = uid();
 
-			const canEditAtDraggedRange = this.isEnabled && editor.model.canEditAt( this._draggedRange );
-
-			data.dataTransfer.effectAllowed = canEditAtDraggedRange ? 'copyMove' : 'copy';
+			data.dataTransfer.effectAllowed = this.isEnabled ? 'copyMove' : 'copy';
 			data.dataTransfer.setData( 'application/ckeditor5-dragging-uid', this._draggingUid );
 
 			const draggedSelection = model.createSelection( this._draggedRange.toRange() );
@@ -312,7 +297,14 @@ export default class DragDrop extends Plugin {
 				method: 'dragstart'
 			} );
 
-			if ( !canEditAtDraggedRange ) {
+			const { dataTransfer, domTarget, domEvent } = data;
+			const { clientX } = domEvent;
+
+			this._updatePreview( { dataTransfer, domTarget, clientX } );
+
+			data.stopPropagation();
+
+			if ( !this.isEnabled ) {
 				this._draggedRange.detach();
 				this._draggedRange = null;
 				this._draggingUid = '';
@@ -325,6 +317,11 @@ export default class DragDrop extends Plugin {
 		this.listenTo<ViewDocumentDragEndEvent>( viewDocument, 'dragend', ( evt, data ) => {
 			this._finalizeDragging( !data.dataTransfer.isCanceled && data.dataTransfer.dropEffect == 'move' );
 		}, { priority: 'low' } );
+
+		// Reset block dragging mode even if dropped outside the editable.
+		this._domEmitter.listenTo( global.document, 'dragend', () => {
+			this._blockMode = false;
+		}, { useCapture: true } );
 
 		// Dragging over the editable.
 		this.listenTo<ViewDocumentDragEnterEvent>( viewDocument, 'dragenter', () => {
@@ -339,7 +336,7 @@ export default class DragDrop extends Plugin {
 		this.listenTo<ViewDocumentDragLeaveEvent>( viewDocument, 'dragleave', () => {
 			// We do not know if the mouse left the editor or just some element in it, so let us wait a few milliseconds
 			// to check if 'dragover' is not fired.
-			this._removeDropMarkerDelayed();
+			dragDropTarget.removeDropMarkerDelayed();
 		} );
 
 		// Handler for moving dragged content over the target area.
@@ -350,16 +347,9 @@ export default class DragDrop extends Plugin {
 				return;
 			}
 
-			this._removeDropMarkerDelayed.cancel();
+			const { clientX, clientY } = ( data as DomEventData<DragEvent> ).domEvent;
 
-			const targetRange = findDropTargetRange( editor, data.targetRanges, data.target );
-
-			// Do not drop if target place is not editable.
-			if ( !editor.model.canEditAt( targetRange ) ) {
-				data.dataTransfer.dropEffect = 'none';
-
-				return;
-			}
+			dragDropTarget.updateDropMarker( data.target, data.targetRanges, clientX, clientY, this._blockMode );
 
 			// If this is content being dragged from another editor, moving out of current editor instance
 			// is not possible until 'dragend' event case will be fixed.
@@ -376,10 +366,7 @@ export default class DragDrop extends Plugin {
 				}
 			}
 
-			/* istanbul ignore else -- @preserve */
-			if ( targetRange ) {
-				this._updateDropMarkerThrottled( targetRange );
-			}
+			evt.stop();
 		}, { priority: 'low' } );
 	}
 
@@ -390,6 +377,7 @@ export default class DragDrop extends Plugin {
 		const editor = this.editor;
 		const view = editor.editing.view;
 		const viewDocument = view.document;
+		const dragDropTarget = editor.plugins.get( DragDropTarget );
 
 		// Update the event target ranges and abort dropping if dropping over itself.
 		this.listenTo<ViewDocumentClipboardInputEvent>( viewDocument, 'clipboardInput', ( evt, data ) => {
@@ -397,14 +385,10 @@ export default class DragDrop extends Plugin {
 				return;
 			}
 
-			const targetRange = findDropTargetRange( editor, data.targetRanges, data.target );
+			const { clientX, clientY } = ( data as DomEventData<DragEvent> ).domEvent;
+			const targetRange = dragDropTarget.getFinalDropRange( data.target, data.targetRanges, clientX, clientY, this._blockMode );
 
-			// The dragging markers must be removed after searching for the target range because sometimes
-			// the target lands on the marker itself.
-			this._removeDropMarker();
-
-			/* istanbul ignore if -- @preserve */
-			if ( !targetRange || !editor.model.canEditAt( targetRange ) ) {
+			if ( !targetRange ) {
 				this._finalizeDragging( false );
 				evt.stop();
 
@@ -499,15 +483,11 @@ export default class DragDrop extends Plugin {
 			// In Firefox this is not needed. In Safari it makes the whole editable draggable (not just textual content).
 			// Disabled in read-only mode because draggable="true" + contenteditable="false" results
 			// in not firing selectionchange event ever, which makes the selection stuck in read-only mode.
-			if ( env.isBlink && !draggableElement && !viewDocument.selection.isCollapsed ) {
+			if ( env.isBlink && !editor.isReadOnly && !draggableElement && !viewDocument.selection.isCollapsed ) {
 				const selectedElement = viewDocument.selection.getSelectedElement();
 
 				if ( !selectedElement || !isWidget( selectedElement ) ) {
-					const editableElement = viewDocument.selection.editableElement;
-
-					if ( editableElement && !editableElement.isReadOnly ) {
-						draggableElement = editableElement;
-					}
+					draggableElement = viewDocument.selection.editableElement;
 				}
 			}
 
@@ -546,82 +526,6 @@ export default class DragDrop extends Plugin {
 	}
 
 	/**
-	 * Creates downcast conversion for the drop target marker.
-	 */
-	private _setupDropMarker(): void {
-		const editor = this.editor;
-
-		// Drop marker conversion for hovering over widgets.
-		editor.conversion.for( 'editingDowncast' ).markerToHighlight( {
-			model: 'drop-target',
-			view: {
-				classes: [ 'ck-clipboard-drop-target-range' ]
-			}
-		} );
-
-		// Drop marker conversion for in text drop target.
-		editor.conversion.for( 'editingDowncast' ).markerToElement( {
-			model: 'drop-target',
-			view: ( data, { writer } ) => {
-				const inText = editor.model.schema.checkChild( data.markerRange.start, '$text' );
-
-				if ( !inText ) {
-					return;
-				}
-
-				return writer.createUIElement( 'span', { class: 'ck ck-clipboard-drop-target-position' }, function( domDocument ) {
-					const domElement = this.toDomElement( domDocument );
-
-					// Using word joiner to make this marker as high as text and also making text not break on marker.
-					domElement.append( '\u2060', domDocument.createElement( 'span' ), '\u2060' );
-
-					return domElement;
-				} );
-			}
-		} );
-	}
-
-	/**
-	 * Updates the drop target marker to the provided range.
-	 *
-	 * @param targetRange The range to set the marker to.
-	 */
-	private _updateDropMarker( targetRange: Range ): void {
-		const editor = this.editor;
-		const markers = editor.model.markers;
-
-		editor.model.change( writer => {
-			if ( markers.has( 'drop-target' ) ) {
-				if ( !markers.get( 'drop-target' )!.getRange().isEqual( targetRange ) ) {
-					writer.updateMarker( 'drop-target', { range: targetRange } );
-				}
-			} else {
-				writer.addMarker( 'drop-target', {
-					range: targetRange,
-					usingOperation: false,
-					affectsData: false
-				} );
-			}
-		} );
-	}
-
-	/**
-	 * Removes the drop target marker.
-	 */
-	private _removeDropMarker(): void {
-		const model = this.editor.model;
-
-		this._removeDropMarkerDelayed.cancel();
-		this._updateDropMarkerThrottled.cancel();
-
-		if ( model.markers.has( 'drop-target' ) ) {
-			model.change( writer => {
-				writer.removeMarker( 'drop-target' );
-			} );
-		}
-	}
-
-	/**
 	 * Deletes the dragged content from its original range and clears the dragging state.
 	 *
 	 * @param moved Whether the move succeeded.
@@ -629,8 +533,9 @@ export default class DragDrop extends Plugin {
 	private _finalizeDragging( moved: boolean ): void {
 		const editor = this.editor;
 		const model = editor.model;
+		const dragDropTarget = editor.plugins.get( DragDropTarget );
 
-		this._removeDropMarker();
+		dragDropTarget.removeDropMarker();
 		this._clearDraggableAttributes();
 
 		if ( editor.plugins.has( 'WidgetToolbarRepository' ) ) {
@@ -640,6 +545,11 @@ export default class DragDrop extends Plugin {
 		}
 
 		this._draggingUid = '';
+
+		if ( this._previewContainer ) {
+			this._previewContainer.remove();
+			this._previewContainer = undefined;
+		}
 
 		if ( !this._draggedRange ) {
 			return;
@@ -653,171 +563,112 @@ export default class DragDrop extends Plugin {
 		this._draggedRange.detach();
 		this._draggedRange = null;
 	}
-}
 
-/**
- * Returns fixed selection range for given position and target element.
- */
-function findDropTargetRange( editor: Editor, targetViewRanges: Array<ViewRange> | null, targetViewElement: ViewElement ): Range | null {
-	const model = editor.model;
-	const mapper = editor.editing.mapper;
+	/**
+	 * Sets the dragged source range based on event target and document selection.
+	 */
+	private _prepareDraggedRange( target: ViewElement ): void {
+		const editor = this.editor;
+		const model = editor.model;
+		const selection = model.document.selection;
 
-	let range: Range | null = null;
+		// Check if this is dragstart over the widget (but not a nested editable).
+		const draggableWidget = target ? findDraggableWidget( target ) : null;
 
-	const targetViewPosition = targetViewRanges ? targetViewRanges[ 0 ].start : null;
+		if ( draggableWidget ) {
+			const modelElement = editor.editing.mapper.toModelElement( draggableWidget )!;
 
-	// A UIElement is not a valid drop element, use parent (this could be a drop marker or any other UIElement).
-	if ( targetViewElement.is( 'uiElement' ) ) {
-		targetViewElement = targetViewElement.parent as ViewElement;
-	}
+			this._draggedRange = LiveRange.fromRange( model.createRangeOn( modelElement ) );
+			this._blockMode = model.schema.isBlock( modelElement );
 
-	// Quick win if the target is a widget (but not a nested editable).
-	range = findDropTargetRangeOnWidget( editor, targetViewElement );
+			// Disable toolbars so they won't obscure the drop area.
+			if ( editor.plugins.has( 'WidgetToolbarRepository' ) ) {
+				const widgetToolbarRepository: WidgetToolbarRepository = editor.plugins.get( 'WidgetToolbarRepository' );
 
-	if ( range ) {
-		return range;
-	}
+				widgetToolbarRepository.forceDisabled( 'dragDrop' );
+			}
 
-	// The easiest part is over, now we need to move to the model space.
-
-	// Find target model element and position.
-	const targetModelElement = getClosestMappedModelElement( editor, targetViewElement );
-	const targetModelPosition = targetViewPosition ? mapper.toModelPosition( targetViewPosition ) : null;
-
-	// There is no target position while hovering over an empty table cell.
-	// In Safari, target position can be empty while hovering over a widget (e.g., a page-break).
-	// Find the drop position inside the element.
-	if ( !targetModelPosition ) {
-		return findDropTargetRangeInElement( editor, targetModelElement );
-	}
-
-	// Check if target position is between blocks and adjust drop position to the next object.
-	// This is because while hovering over a root element next to a widget the target position can jump in crazy places.
-	range = findDropTargetRangeBetweenBlocks( editor, targetModelPosition, targetModelElement );
-
-	if ( range ) {
-		return range;
-	}
-
-	// Try fixing selection position.
-	// In Firefox, the target position lands before widgets but in other browsers it tends to land after a widget.
-	range = model.schema.getNearestSelectionRange( targetModelPosition, env.isGecko ? 'forward' : 'backward' );
-
-	if ( range ) {
-		return range;
-	}
-
-	// There is no valid selection position inside the current limit element so find a closest object ancestor.
-	// This happens if the model position lands directly in the <table> element itself (view target element was a `<td>`
-	// so a nested editable, but view target position was directly in the `<figure>` element).
-	return findDropTargetRangeOnAncestorObject( editor, targetModelPosition.parent as Element );
-}
-
-/**
- * Returns fixed selection range for a given position and a target element if it is over the widget but not over its nested editable.
- */
-function findDropTargetRangeOnWidget( editor: Editor, targetViewElement: ViewElement ): Range | null {
-	const model = editor.model;
-	const mapper = editor.editing.mapper;
-
-	// Quick win if the target is a widget.
-	if ( isWidget( targetViewElement ) ) {
-		return model.createRangeOn( mapper.toModelElement( targetViewElement )! );
-	}
-
-	// Check if we are deeper over a widget (but not over a nested editable).
-	if ( !targetViewElement.is( 'editableElement' ) ) {
-		// Find a closest ancestor that is either a widget or an editable element...
-		const ancestor = targetViewElement.findAncestor( node => isWidget( node ) || node.is( 'editableElement' ) )!;
-
-		// ...and if the widget was closer then it is a drop target.
-		if ( isWidget( ancestor ) ) {
-			return model.createRangeOn( mapper.toModelElement( ancestor! )! );
-		}
-	}
-
-	return null;
-}
-
-/**
- * Returns fixed selection range inside a model element.
- */
-function findDropTargetRangeInElement( editor: Editor, targetModelElement: Element ): Range | null {
-	const model = editor.model;
-	const schema = model.schema;
-
-	const positionAtElementStart = model.createPositionAt( targetModelElement, 0 );
-
-	return schema.getNearestSelectionRange( positionAtElementStart, 'forward' );
-}
-
-/**
- * Returns fixed selection range for a given position and a target element if the drop is between blocks.
- */
-function findDropTargetRangeBetweenBlocks( editor: Editor, targetModelPosition: Position, targetModelElement: Element ): Range | null {
-	const model = editor.model;
-
-	// Check if target is between blocks.
-	if ( !model.schema.checkChild( targetModelElement, '$block' ) ) {
-		return null;
-	}
-
-	// Find position between blocks.
-	const positionAtElementStart = model.createPositionAt( targetModelElement, 0 );
-
-	// Get the common part of the path (inside the target element and the target position).
-	const commonPath = targetModelPosition.path.slice( 0, positionAtElementStart.path.length );
-
-	// Position between the blocks.
-	const betweenBlocksPosition = model.createPositionFromPath( targetModelPosition.root, commonPath );
-	const nodeAfter = betweenBlocksPosition.nodeAfter;
-
-	// Adjust drop position to the next object.
-	// This is because while hovering over a root element next to a widget the target position can jump in crazy places.
-	if ( nodeAfter && model.schema.isObject( nodeAfter ) ) {
-		return model.createRangeOn( nodeAfter );
-	}
-
-	return null;
-}
-
-/**
- * Returns a selection range on the ancestor object.
- */
-function findDropTargetRangeOnAncestorObject( editor: Editor, element: Element ): Range | null {
-	const model = editor.model;
-	let currentElement: Element | null = element;
-
-	while ( currentElement ) {
-		if ( model.schema.isObject( currentElement ) ) {
-			return model.createRangeOn( currentElement );
+			return;
 		}
 
-		currentElement = currentElement.parent as Element | null;
+		// If this was not a widget we should check if we need to drag some text content.
+		if ( selection.isCollapsed && !( selection.getFirstPosition()!.parent as Element ).isEmpty ) {
+			return;
+		}
+
+		const blocks = Array.from( selection.getSelectedBlocks() );
+		const draggedRange = selection.getFirstRange()!;
+
+		if ( blocks.length == 0 ) {
+			this._draggedRange = LiveRange.fromRange( draggedRange );
+
+			return;
+		}
+
+		const blockRange = getRangeIncludingFullySelectedParents( model, blocks );
+
+		if ( blocks.length > 1 ) {
+			this._draggedRange = LiveRange.fromRange( blockRange );
+			this._blockMode = true;
+			// TODO block mode for dragging from outside editor? or inline? or both?
+		} else if ( blocks.length == 1 ) {
+			const touchesBlockEdges = draggedRange.start.isTouching( blockRange.start ) &&
+					draggedRange.end.isTouching( blockRange.end );
+
+			this._draggedRange = LiveRange.fromRange( touchesBlockEdges ? blockRange : draggedRange );
+			this._blockMode = touchesBlockEdges;
+		}
+
+		model.change( writer => writer.setSelection( this._draggedRange!.toRange() ) );
 	}
 
-	/* istanbul ignore next -- @preserve */
-	return null;
-}
+	/**
+	 * Updates the dragged preview image.
+	 */
+	private _updatePreview( {
+		dataTransfer,
+		domTarget,
+		clientX
+	}: {
+		dataTransfer: DataTransfer;
+		domTarget: HTMLElement;
+		clientX: number;
+	} ): void {
+		const view = this.editor.editing.view;
+		const editable = view.document.selection.editableElement!;
+		const domEditable = view.domConverter.mapViewToDom( editable )!;
+		const computedStyle = global.window.getComputedStyle( domEditable );
 
-/**
- * Returns the closest model element for the specified view element.
- */
-function getClosestMappedModelElement( editor: Editor, element: ViewElement ): Element {
-	const mapper = editor.editing.mapper;
-	const view = editor.editing.view;
+		if ( !this._previewContainer ) {
+			this._previewContainer = createElement( global.document, 'div', {
+				style: 'position: fixed; left: -999999px;'
+			} );
 
-	const targetModelElement = mapper.toModelElement( element );
+			global.document.body.appendChild( this._previewContainer );
+		} else if ( this._previewContainer.firstElementChild ) {
+			this._previewContainer.removeChild( this._previewContainer.firstElementChild );
+		}
 
-	if ( targetModelElement ) {
-		return targetModelElement;
+		const domRect = new Rect( domEditable );
+
+		// If domTarget is inside the editable root, browsers will display the preview correctly by themselves.
+		if ( domEditable.contains( domTarget ) ) {
+			return;
+		}
+
+		const domEditablePaddingLeft = parseFloat( computedStyle.paddingLeft );
+		const preview = createElement( global.document, 'div' );
+
+		preview.className = 'ck ck-content';
+		preview.style.width = computedStyle.width;
+		preview.style.paddingLeft = `${ domRect.left - clientX + domEditablePaddingLeft }px`;
+
+		preview.innerHTML = dataTransfer.getData( 'text/html' );
+
+		dataTransfer.setDragImage( preview, 0, 0 );
+
+		this._previewContainer.appendChild( preview );
 	}
-
-	// Find mapped ancestor if the target is inside not mapped element (for example inline code element).
-	const viewPosition = view.createPositionBefore( element );
-	const viewElement = mapper.findMappedViewAncestor( viewPosition );
-
-	return mapper.toModelElement( viewElement )!;
 }
 
 /**
@@ -860,4 +711,44 @@ function findDraggableWidget( target: ViewElement ): ViewElement | null {
 	}
 
 	return null;
+}
+
+/**
+ * Recursively checks if common parent of provided elements doesn't have any other children. If that's the case,
+ * it returns range including this parent. Otherwise, it returns only the range from first to last element.
+ *
+ * Example:
+ *
+ * <blockQuote>
+ *   <paragraph>[Test 1</paragraph>
+ *   <paragraph>Test 2</paragraph>
+ *   <paragraph>Test 3]</paragraph>
+ * <blockQuote>
+ *
+ * Because all elements inside the `blockQuote` are selected, the range is extended to include the `blockQuote` too.
+ * If only first and second paragraphs would be selected, the range would not include it.
+ */
+function getRangeIncludingFullySelectedParents( model: Model, elements: Array<Element> ): Range {
+	const firstElement = elements[ 0 ];
+	const lastElement = elements[ elements.length - 1 ];
+	const parent = firstElement.getCommonAncestor( lastElement );
+	const startPosition: Position = model.createPositionBefore( firstElement );
+	const endPosition: Position = model.createPositionAfter( lastElement );
+
+	if (
+		parent &&
+		parent.is( 'element' ) &&
+		!model.schema.isLimit( parent )
+	) {
+		const parentRange = model.createRangeOn( parent );
+		const touchesStart = startPosition.isTouching( parentRange.start );
+		const touchesEnd = endPosition.isTouching( parentRange.end );
+
+		if ( touchesStart && touchesEnd ) {
+			// Selection includes all elements in the parent.
+			return getRangeIncludingFullySelectedParents( model, [ parent ] );
+		}
+	}
+
+	return model.createRange( startPosition, endPosition );
 }
