+++ packages/ckeditor5-clipboard/src/index.ts
@@ -22,7 +22,6 @@ export type {
 
 export { default as DragDrop } from './dragdrop';
 export { default as PastePlainText } from './pasteplaintext';
-export { default as DragDropExperimental } from './dragdropexperimental';
 export { default as DragDropTarget } from './dragdroptarget';
 export { default as DragDropBlockToolbar } from './dragdropblocktoolbar';
