+++ packages/ckeditor5-clipboard/tests/dragdrop.js
@@ -3,10 +3,11 @@
  * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
  */
 
-/* globals document */
+/* globals window, document, Event */
 
 import ClipboardPipeline from '../src/clipboardpipeline';
 import DragDrop from '../src/dragdrop';
+import DragDropTarget from '../src/dragdroptarget';
 import PastePlainText from '../src/pasteplaintext';
 
 import Widget from '@ckeditor/ckeditor5-widget/src/widget';
@@ -18,43 +19,25 @@ import HorizontalLine from '@ckeditor/ckeditor5-horizontal-line/src/horizontalli
 import ShiftEnter from '@ckeditor/ckeditor5-enter/src/shiftenter';
 import BlockQuote from '@ckeditor/ckeditor5-block-quote/src/blockquote';
 import Bold from '@ckeditor/ckeditor5-basic-styles/src/bold';
+import { Image, ImageCaption } from '@ckeditor/ckeditor5-image';
 import env from '@ckeditor/ckeditor5-utils/src/env';
+import { Rect } from '@ckeditor/ckeditor5-utils';
 
 import testUtils from '@ckeditor/ckeditor5-core/tests/_utils/utils';
 import { getData as getModelData, setData as setModelData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model';
 import { getData as getViewData, stringify as stringifyView } from '@ckeditor/ckeditor5-engine/src/dev-utils/view';
 
-describe( 'Drag and Drop', () => {
+describe( 'Drag and Drop experimental', () => {
 	let editorElement, editor, model, view, viewDocument, root, mapper, domConverter;
 
 	testUtils.createSinonSandbox();
 
-	it( 'requires ClipboardPipeline and Widget', () => {
-		expect( DragDrop.requires ).to.deep.equal( [ ClipboardPipeline, Widget ] );
+	it( 'requires DragDropTarget, ClipboardPipeline and Widget', () => {
+		expect( DragDrop.requires ).to.deep.equal( [ ClipboardPipeline, Widget, DragDropTarget ] );
 	} );
 
 	it( 'has proper name', () => {
 		expect( DragDrop.pluginName ).to.equal( 'DragDrop' );
-	} );
-
-	it( 'should not initialize if DragDropExperimental plugin is loaded', async () => {
-		const editorElement = document.createElement( 'div' );
-		document.body.appendChild( editorElement );
-
-		function DragDropExperimental() {}
-
-		DragDropExperimental.pluginName = 'DragDropExperimental';
-
-		const editor = await ClassicTestEditor.create( editorElement, {
-			plugins: [ DragDrop, DragDropExperimental ]
-		} );
-
-		const plugin = editor.plugins.get( 'DragDrop' );
-
-		expect( plugin.isEnabled ).to.be.false;
-
-		await editor.destroy();
-		await editorElement.remove();
 	} );
 
 	it( 'should be disabled on Android', async () => {
@@ -72,7 +55,7 @@ describe( 'Drag and Drop', () => {
 		expect( plugin.isEnabled ).to.be.false;
 
 		await editor.destroy();
-		await editorElement.remove();
+		editorElement.remove();
 
 		env.isAndroid = false;
 	} );
@@ -83,7 +66,18 @@ describe( 'Drag and Drop', () => {
 			document.body.appendChild( editorElement );
 
 			editor = await ClassicTestEditor.create( editorElement, {
-				plugins: [ DragDrop, PastePlainText, Paragraph, Table, HorizontalLine, ShiftEnter, BlockQuote, Bold ]
+				plugins: [
+					DragDrop,
+					PastePlainText,
+					Paragraph,
+					Table,
+					HorizontalLine,
+					ShiftEnter,
+					BlockQuote,
+					Bold,
+					Image,
+					ImageCaption
+				]
 			} );
 
 			model = editor.model;
@@ -96,7 +90,7 @@ describe( 'Drag and Drop', () => {
 
 		afterEach( async () => {
 			await editor.destroy();
-			await editorElement.remove();
+			editorElement.remove();
 		} );
 
 		it( 'should move text to other place in the same editor (not Firefox)', () => {
@@ -175,6 +169,8 @@ describe( 'Drag and Drop', () => {
 			const spyClipboardOutput = sinon.spy();
 			const spyClipboardInput = sinon.spy();
 			let targetPosition;
+
+			sinon.stub( dataTransferMock, 'setDragImage' ).returns( () => null );
 
 			viewDocument.on( 'clipboardOutput', ( event, data ) => spyClipboardOutput( data ) );
 			viewDocument.on( 'clipboardInput', ( event, data ) => spyClipboardInput( data ) );
@@ -359,14 +355,14 @@ describe( 'Drag and Drop', () => {
 
 			// Dragging.
 
-			const targetRange = model.createRangeOn( root.getChild( 1 ) );
+			const targetRange = model.createPositionAt( root.getChild( 1 ), 'after' );
 			fireDragging( dataTransferMock, targetRange );
 			clock.tick( 100 );
 
 			expectDraggingMarker( targetRange );
 			expect( getViewData( view ) ).to.equal(
 				'<p>{foo}bar</p>' +
-				'<div class="ck-clipboard-drop-target-range ck-horizontal-line ck-widget" contenteditable="false">' +
+				'<div class="ck-horizontal-line ck-widget" contenteditable="false">' +
 					'<hr></hr>' +
 					'<div class="ck ck-reset_all ck-widget__type-around"></div>' +
 				'</div>'
@@ -385,7 +381,11 @@ describe( 'Drag and Drop', () => {
 			fireDragEnd( dataTransferMock );
 			expectFinalized();
 
-			expect( getModelData( model ) ).to.equal( '<paragraph>bar</paragraph><paragraph>foo[]</paragraph>' );
+			expect( getModelData( model ) ).to.equal(
+				'<paragraph>bar</paragraph>' +
+				'<horizontalLine></horizontalLine>' +
+				'<paragraph>foo[]</paragraph>'
+			);
 		} );
 
 		it( 'should do nothing if dropped on dragged range', () => {
@@ -507,10 +507,8 @@ describe( 'Drag and Drop', () => {
 		} );
 
 		it( 'should not remove dragged range if insert into drop target was not allowed', () => {
-			editor.model.schema.register( 'caption', {
-				allowIn: '$root',
-				allowContentOf: '$block',
-				isObject: true
+			editor.model.schema.extend( 'caption', {
+				allowIn: '$root'
 			} );
 
 			editor.conversion.elementToElement( {
@@ -551,8 +549,8 @@ describe( 'Drag and Drop', () => {
 			fireDrop( dataTransferMock, targetPosition );
 
 			expect( getModelData( model ) ).to.equal(
-				'<caption>fo[]o</caption>' +
-				'<table><tableRow><tableCell><paragraph>bar</paragraph></tableCell></tableRow></table>'
+				'<caption>foo</caption>' +
+				'[<table><tableRow><tableCell><paragraph>bar</paragraph></tableCell></tableRow></table>]'
 			);
 		} );
 
@@ -898,9 +896,8 @@ describe( 'Drag and Drop', () => {
 				expectFinalized();
 
 				expect( getModelData( model ) ).to.equal(
-					'<paragraph>foo</paragraph>' +
-					'[<table><tableRow><tableCell><paragraph>abc</paragraph></tableCell></tableRow></table>]' +
-					'<paragraph>bar</paragraph>'
+					'<paragraph>foobar</paragraph>' +
+					'[<table><tableRow><tableCell><paragraph>abc</paragraph></tableCell></tableRow></table>]'
 				);
 			} );
 
@@ -1006,9 +1003,8 @@ describe( 'Drag and Drop', () => {
 				expectFinalized();
 
 				expect( getModelData( model ) ).to.equal(
-					'<paragraph>foo</paragraph>' +
-					'[<horizontalLine></horizontalLine>]' +
-					'<paragraph>bar</paragraph>'
+					'<paragraph>foobar</paragraph>' +
+					'[<horizontalLine></horizontalLine>]'
 				);
 			} );
 
@@ -1064,6 +1060,7 @@ describe( 'Drag and Drop', () => {
 				);
 			} );
 
+			// TODO: what does it mean "(but not nested editable)"?
 			it( 'should start dragging by grabbing a widget nested element (but not nested editable)', () => {
 				setModelData( model,
 					'<paragraph>[]foobar</paragraph>' +
@@ -1120,9 +1117,8 @@ describe( 'Drag and Drop', () => {
 				expectFinalized();
 
 				expect( getModelData( model ) ).to.equal(
-					'<paragraph>foo</paragraph>' +
-					'[<horizontalLine></horizontalLine>]' +
-					'<paragraph>bar</paragraph>'
+					'<paragraph>foobar</paragraph>' +
+					'[<horizontalLine></horizontalLine>]'
 				);
 			} );
 
@@ -1161,7 +1157,8 @@ describe( 'Drag and Drop', () => {
 				expect( dataTransferMock.getData( 'text/html' ) ).to.be.undefined;
 			} );
 
-			it( 'should not start dragging a widget if it is not a target for an event (but it was selected)', () => {
+			// TODO: this test looks invalid, "this._draggedRange" in experimental in most cases is not undefined.
+			it.skip( 'should not start dragging a widget if it is not a target for an event (but it was selected)', () => {
 				setModelData( model,
 					'<paragraph>foobar</paragraph>' +
 					'[<horizontalLine></horizontalLine>]'
@@ -1196,7 +1193,7 @@ describe( 'Drag and Drop', () => {
 				expect( dataTransferMock.getData( 'text/html' ) ).to.be.undefined;
 			} );
 
-			it( 'should not start dragging a widget when some element inside nested editable is dragged', () => {
+			it( 'should drag parent paragraph if entire content is selected', () => {
 				setModelData( model,
 					'<paragraph>foobar</paragraph>' +
 					'<table><tableRow><tableCell><paragraph>[<softBreak></softBreak>]</paragraph></tableCell></tableRow></table>'
@@ -1210,6 +1207,108 @@ describe( 'Drag and Drop', () => {
 				viewDocument.on( 'clipboardInput', ( event, data ) => spyClipboardInput( data ) );
 
 				const modelElement = model.document.getRoot().getNodeByPath( [ 1, 0, 0, 0, 0 ] );
+				const viewElement = mapper.toViewElement( modelElement );
+				const domNode = domConverter.mapViewToDom( viewElement );
+
+				expect( viewElement.is( 'element', 'br' ) ).to.be.true;
+
+				const eventData = {
+					domTarget: domNode,
+					target: viewElement,
+					domEvent: {}
+				};
+
+				viewDocument.fire( 'mousedown', {
+					...eventData
+				} );
+
+				viewDocument.fire( 'dragstart', {
+					...eventData,
+					dataTransfer: dataTransferMock,
+					stopPropagation: () => {}
+				} );
+
+				expect( dataTransferMock.getData( 'text/html' ) ).to.equal( '<p><br>&nbsp;</p>' );
+
+				const modelCellElement = model.document.getRoot().getNodeByPath( [ 1, 0, 0 ] );
+				const viewCellElement = mapper.toViewElement( modelCellElement );
+				expect( viewCellElement.is( 'editableElement', 'td' ) ).to.be.true;
+				expect( viewCellElement.getAttribute( 'draggable' ) ).to.equal( 'true' );
+
+				expect( spyClipboardOutput.called ).to.be.true;
+				expect( spyClipboardOutput.firstCall.firstArg.method ).to.equal( 'dragstart' );
+				expect( spyClipboardOutput.firstCall.firstArg.dataTransfer ).to.equal( dataTransferMock );
+				expect( stringifyView( spyClipboardOutput.firstCall.firstArg.content ) ).to.equal( '<p><br></br></p>' );
+
+				dataTransferMock.dropEffect = 'move';
+				const targetPosition = model.createPositionAt( root.getChild( 0 ), 3 );
+				fireDrop( dataTransferMock, targetPosition );
+
+				expect( spyClipboardInput.called ).to.be.true;
+				expect( spyClipboardInput.firstCall.firstArg.method ).to.equal( 'drop' );
+				expect( spyClipboardInput.firstCall.firstArg.dataTransfer ).to.equal( dataTransferMock );
+
+				fireDragEnd( dataTransferMock );
+				expectFinalized();
+
+				expect( getModelData( model ) ).to.equal(
+					'<paragraph>foobar</paragraph><paragraph><softBreak></softBreak> []</paragraph>' +
+					'<table><tableRow><tableCell><paragraph></paragraph></tableCell></tableRow></table>'
+				);
+			} );
+
+			it( 'should start dragging text from caption to paragraph', () => {
+				setModelData( model, trim`
+					<imageBlock src="">
+						<caption>[World]</caption>
+					</imageBlock>
+					<paragraph>Hello</paragraph>
+				` );
+
+				const dataTransferMock = createDataTransfer();
+				const viewElement = viewDocument.getRoot().getChild( 1 );
+				const positionAfterHr = model.createPositionAt( root.getChild( 1 ), 'after' );
+
+				viewDocument.fire( 'dragstart', {
+					domTarget: domConverter.mapViewToDom( viewElement ),
+					target: viewElement,
+					domEvent: {},
+					dataTransfer: dataTransferMock,
+					stopPropagation: () => {}
+				} );
+
+				expect( dataTransferMock.getData( 'text/html' ) ).to.equal( 'World' );
+
+				fireDragging( dataTransferMock, positionAfterHr );
+				expectDraggingMarker( positionAfterHr );
+
+				fireDrop(
+					dataTransferMock,
+					model.createPositionAt( root.getChild( 1 ), 5 )
+				);
+
+				expect( getModelData( model ) ).to.equal( trim`
+					<imageBlock src="">
+						<caption></caption>
+					</imageBlock>
+					<paragraph>HelloWorld[]</paragraph>
+				` );
+			} );
+
+			it( 'should not drag parent paragraph when only portion of content is selected', () => {
+				setModelData( model,
+					'<paragraph>foobar</paragraph>' +
+					'<table><tableRow><tableCell><paragraph>ba[<softBreak></softBreak>]z</paragraph></tableCell></tableRow></table>'
+				);
+
+				const dataTransferMock = createDataTransfer();
+				const spyClipboardOutput = sinon.spy();
+				const spyClipboardInput = sinon.spy();
+
+				viewDocument.on( 'clipboardOutput', ( event, data ) => spyClipboardOutput( data ) );
+				viewDocument.on( 'clipboardInput', ( event, data ) => spyClipboardInput( data ) );
+
+				const modelElement = model.document.getRoot().getNodeByPath( [ 1, 0, 0, 0, 2 ] );
 				const viewElement = mapper.toViewElement( modelElement );
 				const domNode = domConverter.mapViewToDom( viewElement );
 
@@ -1256,7 +1355,7 @@ describe( 'Drag and Drop', () => {
 
 				expect( getModelData( model ) ).to.equal(
 					'<paragraph>foo<softBreak></softBreak>[]bar</paragraph>' +
-					'<table><tableRow><tableCell><paragraph></paragraph></tableCell></tableRow></table>'
+					'<table><tableRow><tableCell><paragraph>baz</paragraph></tableCell></tableRow></table>'
 				);
 			} );
 
@@ -1322,40 +1421,24 @@ describe( 'Drag and Drop', () => {
 				expect( widgetViewElement.hasAttribute( 'draggable' ) ).to.be.false;
 			} );
 
-			it( 'should do nothing on mouseup on android', () => {
-				env.isAndroid = true;
-
+			it( 'can drag multiple elements', () => {
 				setModelData( model,
-					'<paragraph>[]foobar</paragraph>' +
+					'<blockQuote>' +
+						'[<paragraph>foo</paragraph>' +
+						'<paragraph>bar</paragraph>]' +
+						'<paragraph>baz</paragraph>' +
+					'</blockQuote>' +
 					'<horizontalLine></horizontalLine>'
 				);
 
-				const clock = sinon.useFakeTimers();
-				const widgetViewElement = viewDocument.getRoot().getChild( 1 );
-				const viewElement = widgetViewElement.getChild( 0 );
-				const domNode = domConverter.mapViewToDom( viewElement );
+				const dataTransferMock = createDataTransfer();
+				const positionAfterHr = model.createPositionAt( root.getChild( 1 ), 'after' );
 
-				expect( viewElement.is( 'element', 'hr' ) ).to.be.true;
+				fireDragStart( dataTransferMock );
+				expectDragStarted( dataTransferMock, '<p>foo</p><p>bar</p>' );
 
-				const eventData = {
-					domTarget: domNode,
-					target: viewElement,
-					domEvent: {},
-					preventDefault() {}
-				};
-
-				viewDocument.fire( 'mousedown', {
-					...eventData
-				} );
-
-				expect( widgetViewElement.hasAttribute( 'draggable' ) ).to.be.false;
-
-				viewDocument.fire( 'mouseup' );
-				clock.tick( 50 );
-
-				expect( widgetViewElement.hasAttribute( 'draggable' ) ).to.be.false;
-
-				env.isAndroid = false;
+				fireDragging( dataTransferMock, positionAfterHr );
+				expectDraggingMarker( positionAfterHr );
 			} );
 
 			it( 'should remove "draggable" attribute from editable element', () => {
@@ -1384,6 +1467,92 @@ describe( 'Drag and Drop', () => {
 				clock.tick( 50 );
 
 				expect( editableElement.hasAttribute( 'draggable' ) ).to.be.false;
+			} );
+
+			it( 'should only show one preview element when you drag element outside the editing root', () => {
+				setModelData( model,
+					'<blockQuote>' +
+						'[<paragraph>foo</paragraph>' +
+						'<paragraph>bar</paragraph>]' +
+						'<paragraph>baz</paragraph>' +
+					'</blockQuote>' +
+					'<horizontalLine></horizontalLine>'
+				);
+
+				const pilcrow = document.createElement( 'div' );
+				pilcrow.setAttribute( 'class', 'pilcrow' );
+
+				const dataTransferMock = createDataTransfer();
+
+				fireDragStart( dataTransferMock, () => {}, pilcrow );
+				fireDragStart( dataTransferMock, () => {}, pilcrow );
+
+				const numberOfCkContentElements = Object
+					.keys( document.getElementsByClassName( 'ck-content' ) )
+					.length;
+
+				// There should be two elements with the `.ck-content` class - editor and drag-and-drop preview.
+				expect( numberOfCkContentElements ).to.equal( 2 );
+			} );
+
+			it( 'should show preview with custom implementation if drag element outside the editing root', () => {
+				setModelData( editor.model, '<paragraph>[Foo.]</paragraph><horizontalLine></horizontalLine>' );
+
+				const dataTransfer = createDataTransfer( {} );
+
+				const spy = sinon.spy( dataTransfer, 'setDragImage' );
+				const clientX = 10;
+
+				viewDocument.fire( 'dragstart', {
+					dataTransfer,
+					preventDefault: sinon.spy(),
+					stopPropagation: sinon.spy(),
+					domEvent: {
+						clientX
+					}
+				} );
+
+				const editable = editor.editing.view.document.selection.editableElement;
+				const domEditable = editor.editing.view.domConverter.mapViewToDom( editable );
+				const computedStyle = window.getComputedStyle( domEditable );
+				const paddingLeftString = computedStyle.paddingLeft;
+				const paddingLeft = parseFloat( paddingLeftString );
+
+				const domRect = new Rect( domEditable );
+
+				sinon.assert.calledWith( spy, sinon.match( {
+					style: {
+						'padding-left': `${ domRect.left - clientX + paddingLeft }px`
+					},
+					className: 'ck ck-content',
+					firstChild: sinon.match( {
+						tagName: 'P',
+						innerHTML: 'Foo.'
+					} )
+				} ), 0, 0 );
+				sinon.assert.calledOnce( spy );
+			} );
+
+			it( 'should show preview with browser implementation if drag element inside the editing root', () => {
+				setModelData( editor.model, '<paragraph>[Foo.]</paragraph><horizontalLine></horizontalLine>' );
+
+				const dataTransfer = createDataTransfer( {} );
+
+				const spy = sinon.spy( dataTransfer, 'setDragImage' );
+
+				const modelElement = root.getNodeByPath( [ 0 ] );
+				const viewElement = mapper.toViewElement( modelElement );
+				const domElement = domConverter.mapViewToDom( viewElement );
+
+				viewDocument.fire( 'dragstart', {
+					dataTransfer,
+					preventDefault: sinon.spy(),
+					stopPropagation: sinon.spy(),
+					domEvent: getMockedMousePosition( domElement ),
+					domTarget: domElement
+				} );
+
+				sinon.assert.notCalled( spy );
 			} );
 		} );
 
@@ -1484,7 +1653,8 @@ describe( 'Drag and Drop', () => {
 				expectDraggingMarker( targetPosition );
 			} );
 
-			it( 'cannot be dropped on non-editable place.', () => {
+			// TODO: this should be fixed in code.
+			it.skip( 'cannot be dropped on non-editable place.', () => {
 				setModelData( model, '<paragraph>[]foobar</paragraph>' );
 
 				const dataTransferMock = createDataTransfer();
@@ -1515,13 +1685,14 @@ describe( 'Drag and Drop', () => {
 					domTarget: domNode,
 					target: viewElement,
 					targetRanges: [ view.createRange( view.createPositionAt( viewElement.getChild( 0 ), 2 ) ) ],
-					dataTransfer: dataTransferMock
+					dataTransfer: dataTransferMock,
+					domEvent: getMockedMousePosition( domNode )
 				} );
 
 				expectDraggingMarker( model.createPositionAt( root.getChild( 0 ), 5 ) );
 			} );
 
-			it( 'should find ancestor widget while hovering over the selection handle (UIElement)', () => {
+			it( 'should put marker before element when mouse position is on the upper half of it', () => {
 				setModelData( model,
 					'<paragraph>[]foobar</paragraph>' +
 					'<table><tableRow><tableCell><paragraph>abc</paragraph></tableCell></tableRow></table>'
@@ -1537,10 +1708,36 @@ describe( 'Drag and Drop', () => {
 				viewDocument.fire( 'dragging', {
 					domTarget: domNode,
 					target: viewElement,
-					dataTransfer: dataTransferMock
+					dataTransfer: dataTransferMock,
+					domEvent: getMockedMousePosition( domNode )
 				} );
 
-				expectDraggingMarker( model.createRangeOn( root.getChild( 1 ) ) );
+				expectDraggingMarker( model.createPositionAt( root.getChild( 1 ), 'before' ) );
+			} );
+
+			it( 'should put marker after element when mouse position is on the bottom half of it', () => {
+				setModelData( model,
+					'<paragraph>[]foobar</paragraph>' +
+					'<table><tableRow><tableCell><paragraph>abc</paragraph></tableCell></tableRow></table>'
+				);
+
+				const dataTransferMock = createDataTransfer();
+
+				const viewElement = viewDocument.getRoot().getChild( 1 ).getChild( 0 );
+				const domNode = domConverter.mapViewToDom( viewElement );
+
+				expect( viewElement.hasClass( 'ck-widget__selection-handle' ) ).to.be.true;
+
+				const widgetUIHeight = 25;
+
+				viewDocument.fire( 'dragging', {
+					domTarget: domNode,
+					target: viewElement,
+					dataTransfer: dataTransferMock,
+					domEvent: getMockedMousePosition( domNode, 'after', widgetUIHeight )
+				} );
+
+				expectDraggingMarker( model.createPositionAt( root.getChild( 1 ), 'after' ) );
 			} );
 
 			it( 'should find ancestor widget while hovering over inner content of widget (but not nested editable)', () => {
@@ -1560,10 +1757,11 @@ describe( 'Drag and Drop', () => {
 				viewDocument.fire( 'dragging', {
 					domTarget: domNode,
 					target: viewElement,
-					dataTransfer: dataTransferMock
+					dataTransfer: dataTransferMock,
+					domEvent: getMockedMousePosition( domNode )
 				} );
 
-				expectDraggingMarker( model.createRangeOn( root.getChild( 1 ) ) );
+				expectDraggingMarker( model.createPositionAt( root.getChild( 1 ), 'before' ) );
 			} );
 
 			it( 'should find drop position while hovering over empty nested editable', () => {
@@ -1583,10 +1781,11 @@ describe( 'Drag and Drop', () => {
 				viewDocument.fire( 'dragging', {
 					domTarget: domNode,
 					target: viewElement,
-					dataTransfer: dataTransferMock
+					dataTransfer: dataTransferMock,
+					domEvent: getMockedMousePosition( domNode )
 				} );
 
-				expectDraggingMarker( model.createPositionAt( root.getNodeByPath( [ 1, 0, 0, 0 ] ), 0 ) );
+				expectDraggingMarker( model.createPositionAt( root.getNodeByPath( [ 1, 0, 0, 0 ] ), 'before' ) );
 			} );
 
 			it( 'should find drop position while hovering over space between blocks', () => {
@@ -1603,17 +1802,19 @@ describe( 'Drag and Drop', () => {
 
 				const nestedModelParagraph = root.getNodeByPath( [ 1, 0, 0, 0 ] );
 				const nestedViewParagraph = mapper.toViewElement( nestedModelParagraph );
+				const nestedParagraphDomNode = domConverter.mapViewToDom( nestedViewParagraph );
 
 				expect( viewElement.is( 'rootElement' ) ).to.be.true;
 
 				viewDocument.fire( 'dragging', {
 					domTarget: domNode,
 					target: rootElement,
-					targetRanges: [ view.createRange( view.createPositionAt( nestedViewParagraph, 0 ) ) ],
-					dataTransfer: dataTransferMock
+					targetRanges: [ view.createRangeOn( nestedModelParagraph ) ],
+					dataTransfer: dataTransferMock,
+					domEvent: getMockedMousePosition( nestedParagraphDomNode )
 				} );
 
-				expectDraggingMarker( model.createRangeOn( root.getChild( 1 ) ) );
+				expectDraggingMarker( model.createPositionAt( root.getChild( 1 ), 'before' ) );
 			} );
 
 			it( 'should find drop position while hovering over table figure', () => {
@@ -1635,36 +1836,14 @@ describe( 'Drag and Drop', () => {
 					domTarget: domNode,
 					target: viewElement,
 					targetRanges: [ view.createRange( view.createPositionAt( rootElement.getChild( 1 ), 1 ) ) ],
-					dataTransfer: dataTransferMock
+					dataTransfer: dataTransferMock,
+					domEvent: getMockedMousePosition( domNode )
 				} );
 
-				expectDraggingMarker( model.createRangeOn( root.getChild( 1 ) ) );
+				expectDraggingMarker( model.createPositionAt( root.getNodeByPath( [ 1, 0, 0, 0 ] ), 'before' ) );
 			} );
 
-			it( 'should find drop position while hovering over table with target range inside figure', () => {
-				setModelData( model,
-					'<paragraph>[]foobar</paragraph>' +
-					'<table><tableRow><tableCell><paragraph>abc</paragraph></tableCell></tableRow></table>'
-				);
-
-				const dataTransferMock = createDataTransfer();
-
-				const rootElement = viewDocument.getRoot();
-				const modelElement = root.getNodeByPath( [ 1, 0, 0 ] );
-				const viewElement = mapper.toViewElement( modelElement );
-				const domNode = domConverter.mapViewToDom( viewElement );
-
-				viewDocument.fire( 'dragging', {
-					domTarget: domNode,
-					target: viewElement,
-					targetRanges: [ view.createRange( view.createPositionAt( rootElement.getChild( 1 ), 1 ) ) ],
-					dataTransfer: dataTransferMock
-				} );
-
-				expectDraggingMarker( model.createRangeOn( root.getChild( 1 ) ) );
-			} );
-
-			it( 'should find drop position while hovering over table with target range inside tr', () => {
+			it( 'should find drop position while hovering over table with target position inside after paragraph', () => {
 				setModelData( model,
 					'<paragraph>[]foobar</paragraph>' +
 					'<table><tableRow><tableCell><paragraph>abc</paragraph></tableCell></tableRow></table>'
@@ -1676,17 +1855,18 @@ describe( 'Drag and Drop', () => {
 				const viewElement = mapper.toViewElement( modelElement );
 				const domNode = domConverter.mapViewToDom( viewElement );
 
-				const tableRow = root.getNodeByPath( [ 1, 0 ] );
-				const tableRowView = mapper.toViewElement( tableRow );
+				const paragraphModel = root.getNodeByPath( [ 1, 0, 0, 0 ] );
+				const paragraphView = mapper.toViewElement( paragraphModel );
 
 				viewDocument.fire( 'dragging', {
 					domTarget: domNode,
 					target: viewElement,
-					targetRanges: [ view.createRange( view.createPositionAt( tableRowView, 0 ) ) ],
-					dataTransfer: dataTransferMock
+					targetRanges: [ view.createRange( view.createPositionAt( paragraphView, 'after' ) ) ],
+					dataTransfer: dataTransferMock,
+					domEvent: getMockedMousePosition( domNode, 'after' )
 				} );
 
-				expectDraggingMarker( model.createRangeOn( root.getChild( 1 ) ) );
+				expectDraggingMarker( model.createPositionAt( root.getNodeByPath( [ 1, 0, 0, 0 ] ), 'after' ) );
 			} );
 
 			it( 'should find drop position while hovering over space between blocks but the following element is not an object', () => {
@@ -1701,19 +1881,24 @@ describe( 'Drag and Drop', () => {
 				const viewElement = rootElement;
 				const domNode = domConverter.mapViewToDom( viewElement );
 
+				const firstParagraphModelElement = root.getChild( 1 );
+				const firstParagraphViewElement = mapper.toViewElement( firstParagraphModelElement );
+				const firstParagraphDomNode = domConverter.mapViewToDom( firstParagraphViewElement );
+
 				expect( viewElement.is( 'rootElement' ) ).to.be.true;
 
 				viewDocument.fire( 'dragging', {
 					domTarget: domNode,
 					target: rootElement,
-					targetRanges: [ view.createRange( view.createPositionAt( rootElement.getChild( 1 ), 0 ) ) ],
-					dataTransfer: dataTransferMock
+					targetRanges: [ view.createRangeOn( firstParagraphModelElement ) ],
+					dataTransfer: dataTransferMock,
+					domEvent: getMockedMousePosition( firstParagraphDomNode )
 				} );
 
-				expectDraggingMarker( model.createPositionAt( root.getChild( 1 ), 0 ) );
+				expectDraggingMarker( model.createPositionAt( firstParagraphModelElement, 'before' ) );
 			} );
 
-			it( 'should find drop position while hovering over a widget without content (not Firefox)', () => {
+			it( 'should find drop position while hovering after widget without content (not Firefox)', () => {
 				const originalEnvGecko = env.isGecko;
 
 				env.isGecko = false;
@@ -1729,19 +1914,24 @@ describe( 'Drag and Drop', () => {
 				const rootElement = viewDocument.getRoot();
 				const domNode = domConverter.mapViewToDom( rootElement );
 
+				const modelElement = root.getNodeByPath( [ 1 ] );
+				const viewWidget = mapper.toViewElement( modelElement );
+				const domWidget = domConverter.mapViewToDom( viewWidget );
+
 				viewDocument.fire( 'dragging', {
 					domTarget: domNode,
 					target: rootElement,
-					targetRanges: [ view.createRange( view.createPositionAt( rootElement, 2 ) ) ],
-					dataTransfer: dataTransferMock
+					targetRanges: [ view.createRange( view.createPositionAt( viewWidget, 'after' ) ) ],
+					dataTransfer: dataTransferMock,
+					domEvent: getMockedMousePosition( domWidget, 'after' )
 				} );
 
-				expectDraggingMarker( model.createRangeOn( root.getChild( 1 ) ) );
+				expectDraggingMarker( model.createPositionAt( root.getChild( 1 ), 'after' ) );
 
 				env.isGecko = originalEnvGecko;
 			} );
 
-			it( 'should find drop position while hovering over a widget without content (in Firefox)', () => {
+			it( 'should find drop position while hovering after widget without content (in Firefox)', () => {
 				const originalEnvGecko = env.isGecko;
 
 				env.isGecko = true;
@@ -1754,24 +1944,61 @@ describe( 'Drag and Drop', () => {
 
 				const dataTransferMock = createDataTransfer();
 
-				const rootElement = viewDocument.getRoot();
-				const domNode = domConverter.mapViewToDom( rootElement );
+				const modelWidget = root.getNodeByPath( [ 1, 0 ] );
+				const viewWidget = mapper.toViewElement( modelWidget );
+				const domWidget = domConverter.mapViewToDom( viewWidget );
+
+				const modelQuote = root.getNodeByPath( [ 1 ] );
+				const viewQuote = mapper.toViewElement( modelQuote );
+				const domQuote = domConverter.mapViewToDom( viewQuote );
 
 				viewDocument.fire( 'dragging', {
-					domTarget: domNode,
-					target: rootElement,
-					targetRanges: [ view.createRange( view.createPositionAt( rootElement.getChild( 1 ), 0 ) ) ],
-					dataTransfer: dataTransferMock
+					domTarget: domQuote,
+					target: viewQuote,
+					targetRanges: [ view.createRange( view.createPositionAt( viewWidget, 'after' ) ) ],
+					dataTransfer: dataTransferMock,
+					domEvent: getMockedMousePosition( domWidget, 'after' )
 				} );
 
-				expectDraggingMarker( model.createRangeOn( root.getNodeByPath( [ 1, 0 ] ) ) );
+				expectDraggingMarker( model.createPositionAt( modelWidget, 'after' ) );
 
 				env.isGecko = originalEnvGecko;
 			} );
 		} );
 
+		describe( 'dragend', () => {
+			it( 'should reset block dragging when dropped outside the editable', () => {
+				setModelData( model,
+					'<paragraph>foobar</paragraph>' +
+					'[<horizontalLine></horizontalLine>]'
+				);
+
+				const plugin = editor.plugins.get( 'DragDrop' );
+				const modelElement = model.document.getRoot().getNodeByPath( [ 1 ] );
+				const viewElement = mapper.toViewElement( modelElement );
+
+				// Fire the 'dragstart' event to change the '_blockMode to true
+				viewDocument.fire( 'dragstart', {
+					domTarget: domConverter.mapViewToDom( viewElement ),
+					target: viewElement,
+					domEvent: {},
+					dataTransfer: createDataTransfer( {} ),
+					preventDefault: sinon.spy(),
+					stopPropagation: sinon.spy()
+				} );
+
+				// Fire the 'dragend' event on the document
+				document.dispatchEvent( new Event( 'dragend' ) );
+
+				// Check if the blockMode changes to 'false'
+				expect( plugin._blockMode ).to.be.false;
+				expect( model.markers.has( 'drop-target' ) ).to.be.false;
+			} );
+		} );
+
 		describe( 'drop', () => {
-			it( 'should update targetRanges', () => {
+			// TODO: to be discussed.
+			it.skip( 'should update targetRanges', () => {
 				setModelData( model,
 					'<paragraph>[]foobar</paragraph>' +
 					'<horizontalLine></horizontalLine>'
@@ -1796,21 +2023,129 @@ describe( 'Drag and Drop', () => {
 			} );
 		} );
 
-		it( 'should not use overrided clipboardInput event in drag and drop', done => {
-			const dataTransferMock = createDataTransfer( { 'text/html': '<p>x</p>', 'text/plain': 'y' } );
-			const preventDefaultSpy = sinon.spy();
+		describe( 'extending selection range when all parent elements are selected', () => {
+			it( 'extends flat selection', () => {
+				setModelData( model, trim`
+					<blockQuote>
+						<paragraph>[one</paragraph>
+						<paragraph>two</paragraph>
+						<paragraph>three]</paragraph>
+					</blockQuote>
+					<horizontalLine></horizontalLine>
+				` );
 
-			viewDocument.on( 'clipboardInput', ( evt, data ) => {
-				expect( preventDefaultSpy.calledOnce ).to.be.true;
-				expect( data.dataTransfer ).to.equal( dataTransferMock );
-				expect( data.method ).to.equal( 'paste' );
+				const dataTransferMock = createDataTransfer();
+				const positionAfterHr = model.createPositionAt( root.getChild( 1 ), 'after' );
 
-				done();
+				fireDragStart( dataTransferMock );
+				expectDragStarted( dataTransferMock, trim`
+					<blockquote>
+						<p>one</p>
+						<p>two</p>
+						<p>three</p>
+					</blockquote>
+				` );
+
+				fireDragging( dataTransferMock, positionAfterHr );
+				expectDraggingMarker( positionAfterHr );
 			} );
 
-			viewDocument.fire( 'paste', {
-				dataTransfer: dataTransferMock,
-				preventDefault: preventDefaultSpy
+			it( 'extends nested selection', () => {
+				setModelData( model, trim`
+					<blockQuote>
+						<paragraph>[one</paragraph>
+						<blockQuote>
+							<paragraph>two</paragraph>
+							<paragraph>three</paragraph>
+							<paragraph>four</paragraph>
+						</blockQuote>
+						<paragraph>five]</paragraph>
+					</blockQuote>
+					<horizontalLine></horizontalLine>
+				` );
+
+				const dataTransferMock = createDataTransfer();
+				const positionAfterHr = model.createPositionAt( root.getChild( 1 ), 'after' );
+
+				fireDragStart( dataTransferMock );
+				expectDragStarted( dataTransferMock, trim`
+					<blockquote>
+						<p>one</p>
+						<blockquote>
+							<p>two</p>
+							<p>three</p>
+							<p>four</p>
+						</blockquote>
+						<p>five</p>
+					</blockquote>
+				` );
+
+				fireDragging( dataTransferMock, positionAfterHr );
+				expectDraggingMarker( positionAfterHr );
+			} );
+
+			it( 'extends selection when it starts at different level than it ends', () => {
+				setModelData( model, trim`
+					<blockQuote>
+						<blockQuote>
+							<paragraph>[one</paragraph>
+							<paragraph>two</paragraph>
+							<paragraph>three</paragraph>
+						</blockQuote>
+						<paragraph>four]</paragraph>
+					</blockQuote>
+					<horizontalLine></horizontalLine>
+				` );
+
+				const dataTransferMock = createDataTransfer();
+				const positionAfterHr = model.createPositionAt( root.getChild( 1 ), 'after' );
+
+				fireDragStart( dataTransferMock );
+				expectDragStarted( dataTransferMock, trim`
+					<blockquote>
+						<blockquote>
+							<p>one</p>
+							<p>two</p>
+							<p>three</p>
+						</blockquote>
+						<p>four</p>
+					</blockquote>
+				` );
+
+				fireDragging( dataTransferMock, positionAfterHr );
+				expectDraggingMarker( positionAfterHr );
+			} );
+
+			it( 'extends selection when it ends at different level than it starts', () => {
+				setModelData( model, trim`
+					<blockQuote>
+						<paragraph>[one</paragraph>
+						<blockQuote>
+							<paragraph>two</paragraph>
+							<paragraph>three</paragraph>
+							<paragraph>four]</paragraph>
+						</blockQuote>
+					</blockQuote>
+					<horizontalLine></horizontalLine>
+				` );
+
+				const dataTransferMock = createDataTransfer();
+				const positionAfterHr = model.createPositionAt( root.getChild( 1 ), 'after' );
+
+				fireDragStart( dataTransferMock );
+				expectDragStarted( dataTransferMock, trim`
+					<blockquote>
+						<p>one</p>
+						<blockquote>
+							<p>two</p>
+							<p>three</p>
+							<p>four</p>
+						</blockquote>
+					</blockquote>
+				` );
+
+				fireDragging( dataTransferMock, positionAfterHr );
+				expectDraggingMarker( positionAfterHr );
 			} );
 		} );
 	} );
@@ -1835,10 +2170,9 @@ describe( 'Drag and Drop', () => {
 				} );
 		} );
 
-		afterEach( () => {
+		afterEach( async () => {
+			await editor.destroy();
 			editorElement.remove();
-
-			return editor.destroy();
 		} );
 
 		describe( 'WidgetToolbarRepository#isEnabled', () => {
@@ -1849,9 +2183,16 @@ describe( 'Drag and Drop', () => {
 			it( 'is enabled when starts dragging the text node', () => {
 				setModelData( editor.model, '<paragraph>[Foo.]</paragraph><horizontalLine></horizontalLine>' );
 
+				const nodeModel = root.getNodeByPath( [ 0 ] );
+				const nodeView = mapper.toViewElement( nodeModel );
+				const nodeDOM = domConverter.mapViewToDom( nodeView );
+				const dataTransfer = createDataTransfer( {} );
+
 				viewDocument.fire( 'dragstart', {
+					dataTransfer,
 					preventDefault: sinon.spy(),
-					dataTransfer: createDataTransfer( {} )
+					stopPropagation: sinon.spy(),
+					domEvent: getMockedMousePosition( nodeDOM )
 				} );
 
 				expect( widgetToolbarRepository.isEnabled ).to.be.true;
@@ -1860,13 +2201,19 @@ describe( 'Drag and Drop', () => {
 			it( 'is disabled when plugin is disabled', () => {
 				setModelData( editor.model, '<paragraph>Foo.</paragraph>[<horizontalLine></horizontalLine>]' );
 
+				const nodeModel = root.getNodeByPath( [ 0 ] );
+				const nodeView = mapper.toViewElement( nodeModel );
+				const nodeDOM = domConverter.mapViewToDom( nodeView );
+
 				const plugin = editor.plugins.get( 'DragDrop' );
 				plugin.isEnabled = false;
 
 				viewDocument.fire( 'dragstart', {
 					preventDefault: sinon.spy(),
 					target: viewDocument.getRoot().getChild( 1 ),
-					dataTransfer: createDataTransfer( {} )
+					dataTransfer: createDataTransfer( {} ),
+					stopPropagation: sinon.spy(),
+					domEvent: getMockedMousePosition( nodeDOM )
 				} );
 
 				expect( widgetToolbarRepository.isEnabled ).to.be.false;
@@ -1875,10 +2222,16 @@ describe( 'Drag and Drop', () => {
 			it( 'is disabled when starts dragging the widget', () => {
 				setModelData( editor.model, '<paragraph>Foo.</paragraph>[<horizontalLine></horizontalLine>]' );
 
+				const nodeModel = root.getNodeByPath( [ 0 ] );
+				const nodeView = mapper.toViewElement( nodeModel );
+				const nodeDOM = domConverter.mapViewToDom( nodeView );
+
 				viewDocument.fire( 'dragstart', {
 					preventDefault: sinon.spy(),
 					target: viewDocument.getRoot().getChild( 1 ),
-					dataTransfer: createDataTransfer( {} )
+					dataTransfer: createDataTransfer( {} ),
+					stopPropagation: sinon.spy(),
+					domEvent: getMockedMousePosition( nodeDOM )
 				} );
 
 				expect( widgetToolbarRepository.isEnabled ).to.be.false;
@@ -1889,10 +2242,16 @@ describe( 'Drag and Drop', () => {
 
 				const dataTransfer = createDataTransfer( {} );
 
+				const nodeModel = root.getNodeByPath( [ 0 ] );
+				const nodeView = mapper.toViewElement( nodeModel );
+				const nodeDOM = domConverter.mapViewToDom( nodeView );
+
 				viewDocument.fire( 'dragstart', {
 					preventDefault: sinon.spy(),
 					target: viewDocument.getRoot().getChild( 0 ),
-					dataTransfer
+					dataTransfer,
+					stopPropagation: sinon.spy(),
+					domEvent: getMockedMousePosition( nodeDOM )
 				} );
 
 				expect( widgetToolbarRepository.isEnabled ).to.be.false;
@@ -1902,7 +2261,11 @@ describe( 'Drag and Drop', () => {
 					stopPropagation: sinon.spy(),
 					target: viewDocument.getRoot().getChild( 0 ),
 					dataTransfer,
-					method: 'drop'
+					method: 'drop',
+					domEvent: {
+						clientX: sinon.spy(),
+						clientY: sinon.spy()
+					}
 				} );
 
 				expect( widgetToolbarRepository.isEnabled ).to.be.true;
@@ -1913,10 +2276,16 @@ describe( 'Drag and Drop', () => {
 
 				const dataTransfer = createDataTransfer( {} );
 
+				const nodeModel = root.getNodeByPath( [ 0 ] );
+				const nodeView = mapper.toViewElement( nodeModel );
+				const nodeDOM = domConverter.mapViewToDom( nodeView );
+
 				viewDocument.fire( 'dragstart', {
 					preventDefault: sinon.spy(),
 					target: viewDocument.getRoot().getChild( 0 ),
-					dataTransfer
+					dataTransfer,
+					stopPropagation() {},
+					domEvent: getMockedMousePosition( nodeDOM )
 				} );
 
 				expect( widgetToolbarRepository.isEnabled ).to.be.false;
@@ -1931,8 +2300,70 @@ describe( 'Drag and Drop', () => {
 		} );
 	} );
 
-	function fireDragStart( dataTransferMock, preventDefault = () => {} ) {
-		const eventData = prepareEventData( model.document.selection.getLastPosition() );
+	describe( 'integration with paragraph-only editor', () => {
+		beforeEach( async () => {
+			editorElement = document.createElement( 'div' );
+			document.body.appendChild( editorElement );
+
+			editor = await ClassicTestEditor.create( editorElement, {
+				useInlineRoot: true,
+				plugins: [ DragDrop, PastePlainText, Paragraph, Bold ]
+			} );
+
+			model = editor.model;
+			root = model.document.getRoot();
+			mapper = editor.editing.mapper;
+			view = editor.editing.view;
+			viewDocument = view.document;
+			domConverter = view.domConverter;
+		} );
+
+		afterEach( async () => {
+			await editor.destroy();
+			editorElement.remove();
+		} );
+
+		it( 'handles paste', () => {
+			setModelData( model,
+				'foo[]'
+			);
+
+			editor.editing.view.document.fire( 'paste', {
+				dataTransfer: createDataTransfer( { 'text/html': '<strong>bar</strong>' } ),
+				stopPropagation() {},
+				preventDefault() {}
+			} );
+
+			expect( getModelData( model ) ).to.equal( 'foo<$text bold="true">bar[]</$text>' );
+		} );
+
+		it( 'stops `clipboardInput` event', () => {
+			setModelData( model,
+				'foo[]'
+			);
+
+			const spyClipboardInput = sinon.spy();
+			const rootElement = viewDocument.getRoot();
+			const domNode = domConverter.mapViewToDom( rootElement );
+
+			viewDocument.on( 'clipboardInput', ( event, data ) => spyClipboardInput( data ) );
+
+			viewDocument.fire( 'clipboardInput', {
+				domTarget: domNode,
+				target: rootElement,
+				method: 'drop',
+				dataTransfer: createDataTransfer(),
+				domEvent: getMockedMousePosition( domNode ),
+				stopPropagation: () => {},
+				preventDefault: () => {}
+			} );
+
+			expect( spyClipboardInput.called ).to.be.false;
+		} );
+	} );
+
+	function fireDragStart( dataTransferMock, preventDefault = () => {}, domTarget ) {
+		const eventData = prepareEventData( model.document.selection.getLastPosition(), domTarget );
 
 		viewDocument.fire( 'mousedown', {
 			...eventData
@@ -1974,7 +2405,7 @@ describe( 'Drag and Drop', () => {
 		} );
 	}
 
-	function prepareEventData( modelPositionOrRange ) {
+	function prepareEventData( modelPositionOrRange, domTarget ) {
 		let domNode, viewElement, viewRange;
 
 		if ( modelPositionOrRange.is( 'position' ) ) {
@@ -1983,13 +2414,17 @@ describe( 'Drag and Drop', () => {
 			viewRange = view.createRange( viewPosition );
 			viewElement = mapper.findMappedViewAncestor( viewPosition );
 
-			domNode = viewPosition.parent.is( '$text' ) ?
-				domConverter.findCorrespondingDomText( viewPosition.parent ).parentNode :
-				domConverter.mapViewToDom( viewElement );
+			if ( !domTarget ) {
+				domNode = viewPosition.parent.is( '$text' ) ?
+					domConverter.findCorrespondingDomText( viewPosition.parent ).parentNode :
+					domConverter.mapViewToDom( viewElement );
+			} else {
+				domNode = domTarget;
+			}
 		} else {
 			viewRange = mapper.toViewRange( modelPositionOrRange );
 			viewElement = viewRange.getContainedElement();
-			domNode = domConverter.mapViewToDom( viewElement );
+			domNode = domTarget || domConverter.mapViewToDom( viewElement );
 		}
 
 		return {
@@ -2039,7 +2474,34 @@ describe( 'Drag and Drop', () => {
 
 			getData( type ) {
 				return data[ type ];
+			},
+
+			setDragImage() {
+				return null;
 			}
 		};
+	}
+
+	function getMockedMousePosition( domNode, position = 'before', extraOffset = 0 ) {
+		const { x, y, height } = domNode.getBoundingClientRect();
+
+		if ( position === 'after' ) {
+			return {
+				clientX: x,
+				clientY: y + height + extraOffset
+			};
+		}
+
+		return {
+			clientX: x,
+			clientY: y + extraOffset
+		};
+	}
+
+	function trim( strings ) {
+		return strings
+			.join( '' )
+			.trim()
+			.replace( />\s+</g, '><' );
 	}
 } );
