--- packages/ckeditor5-clipboard/tests/manual/dragdrop.md
@@ -1,7 +1,0 @@
-# Drag & Drop
-
-Main focus of this test is accurate finding the proper drop target.
-
-* Use the draggable elements on the right-side of the editor to drop them to the editor.
-* Move different widgets and text fragments within the editor or outside from it.
-
